just a small typecast to avoid a compiler warning
git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/trunk@465 979fb63a-54b4-4514-8d67-9ab8ea1b59df

+++ src/Index.cpp
@@ -623,7 +623,7 @@ int Index::global_to_stream(void **buffer,size_t *length) {
 
     //Check if we stopped buffering, if so return -1 and length of -1
     if(buffering && buffer_filled){
-        *length=-1;
+        *length=(size_t)-1;
         return -1;
     }
